ajuste no layout de register.py

+++ register.py
@@ -11,13 +11,16 @@ from werkzeug.security import generate_password_hash
 import numpy as np
 import pandas as pd
 
+import os
+
 card_style = {
     'width': '400px',
     'min-height': '300px',
     'padding-top': '25px',
     'padding-right': '25px',
     'padding-left': '25px',
-    'align-self': 'center'
+    'align-self': 'center',
+    'border':'1px solid #ced4da'
 }
 
 button_style = {
@@ -50,6 +53,7 @@ input_style = {
     'border': '1px solid #ccc'
 }
 
+#card_style = {'margin': '20px', 'border': '1px solid #ced4da', 'border-radius': '5px'}
 
 def render_layout(message):
     conn1 = sqlite3.connect(path_db)
@@ -59,22 +63,21 @@ def render_layout(message):
     conn1.close()
     
     message = "Ocorreu um erro durante o registro" if message == "error" else message
-    register = html.Div([
+    register = dbc.Container([
             #dcc.State('df_today_users', data=df.to_dict('records')),
             dbc.Card([
                 dbc.Row([
                     dbc.Col([
                         dbc.Card([
-                            dbc.Row(html.Legend('Registros'), style={'display': 'flex', 'justify-content': 'center'}),
-                            dbc.Row([
-                                dbc.Input(id='user-register', placeholder='Username', type='text', style=input_style),
+                            html.Legend('Registros'),
+
+                            dbc.Input(id='user-register', placeholder='Username', type='text', style=input_style),
                                 #dbc.Input(id='email-register', placeholder='E-mail', type='text'),
-                                dbc.Input(id='pwd-register', placeholder='Password', type='password', style=input_style)
-                            ], style={'display': 'flex', 'justify-content': 'center'}),
+                            dbc.Input(id='pwd-register', placeholder='Password', type='password', style=input_style),
                             dbc.Row([
                                 dbc.Col(dbc.Button('Salvar', id='register-button', style=button_style)),
-                                dbc.Col(dbc.Button('Deletar', id='delete-register-button', style=button_style_cancel))
-                            ], style={'display': 'flex', 'justify-content': 'center'}),
+                                dbc.Col(dbc.Button('Deletar', id='delete-register-button', style=button_style_cancel)),
+                            ], style={'display': 'flex'}),
                             html.Span(message, style={'text-align': 'center'}),
                             html.Div([
                                 html.Label('ou', style={'margin-right': '5px'}),
@@ -83,27 +86,27 @@ def render_layout(message):
                             html.Div(id='clicka')
                         
                         ], style=card_style)
-                    ]),
+                    ], md=6),
                     dbc.Col([
-                        dbc.Card([
-                            dbc.Row([
-                                html.B('Lista de Usuarios')
-                            ]),
-                            dbc.Row([
-                                dash_table.DataTable(
-                                    id='tabela-usuarios',
-                                    columns=[{'name': col, 'id': col} for col in df.columns],
-                                    data=df.to_dict('records'),
-                                    editable=True,
-                                    row_selectable="single",
-                                    selected_rows=[],
-                                )
-                            ], style={'height': '300px', 'overflowY': 'auto'})
-                        ])
-                    ])
-                ])
+                            dbc.Card([
+                                dbc.Row([
+                                    html.B('Lista de Usuarios')
+                                ]),
+                                dbc.Row([
+                                    dash_table.DataTable(
+                                        id='tabela-usuarios',
+                                        columns=[{'name': col, 'id': col} for col in df.columns],
+                                        data=df.to_dict('records'),
+                                        editable=True,
+                                        row_selectable="single",
+                                        selected_rows=[],
+                                    )
+                                ], style={'height': '300px', 'overflowY': 'auto', 'width': '20vw'})
+                            ], style={'padding-left': '10px'})
+                    ], md=6)
+                ], style={'display': 'flex', 'justify-content': 'center'})
             ])
-    ], style={'height': '100vh', 'display': 'flex', 'justify-content': 'center'})
+    ], fluid=True, style={"margin-top": '20vh'})#, style={'height': '100vh', 'display': 'flex', 'justify-content': 'center'})
     
     return register
 
@@ -127,13 +130,17 @@ def register(n_clicks_register, n_clicks_delete, username, password, selected_ro
         if username is not None and password is not None:
             if selected_rows == []:
                 hashed_password = generate_password_hash(password, method='pbkdf2:sha256')
-        
+                
+                username = username.upper()
                 ins = Users_table.insert().values(username=username, password=hashed_password)
                 conn = engine.connect()
         
                 conn.execute(ins)
                 conn.commit()
                 conn.close()
+                
+                if os.path.exists(diretory):
+                    os.makedirs(f"{os.path.abspath(diretory)}\\{username}")
             else:
                 hashed_password = generate_password_hash(password, method='pbkdf2:sha256')
                 
@@ -156,6 +163,8 @@ def register(n_clicks_register, n_clicks_delete, username, password, selected_ro
             id_edit = int(registers[selected_rows[0]]['id'])
             user_to_delete = session.query(Users).filter_by(id=id_edit).first() 
             
+            if user_to_delete.username == 'admin':
+                return "error"#, registers
             session.delete(user_to_delete)
             session.commit()
